+++ examples/svg.rs
@@ -21,12 +21,12 @@ fn main() {
             &current_layer,
             SvgTransform {
                 rotate: Some(SvgRotation {
-                    angle_ccw_degrees: i as f64 * 36.0,
+                    angle_ccw_degrees: i as f32 * 36.0,
                     rotation_center_x: rotation_center_x.into_pt(300.0),
                     rotation_center_y: rotation_center_y.into_pt(300.0),
                 }),
-                translate_x: Some(Mm(i as f64 * 20.0 % 50.0).into()),
-                translate_y: Some(Mm(i as f64 * 30.0).into()),
+                translate_x: Some(Mm(i as f32 * 20.0 % 50.0).into()),
+                translate_y: Some(Mm(i as f32 * 30.0).into()),
                 ..Default::default()
             },
         );
