+++ src/svg.rs
@@ -52,15 +52,15 @@ pub struct SvgTransform {
     pub translate_y: Option<Pt>,
     /// Rotate (clockwise), in degree angles
     pub rotate: Option<SvgRotation>,
-    pub scale_x: Option<f64>,
-    pub scale_y: Option<f64>,
+    pub scale_x: Option<f32>,
+    pub scale_y: Option<f32>,
     /// If set to None, will be set to 300.0 for images
-    pub dpi: Option<f64>,
+    pub dpi: Option<f32>,
 }
 
 #[derive(Debug, Copy, Clone, PartialEq, Default)]
 pub struct SvgRotation {
-    pub angle_ccw_degrees: f64,
+    pub angle_ccw_degrees: f32,
     pub rotation_center_x: Pt,
     pub rotation_center_y: Pt,
 }
